refactor: eliminate code duplication in auth tests
- Extract common test patterns to auth-test-utils.ts
- Add setupEnvironment, testMiddlewareAuth, and verifyUrlValidation helpers
- Remove duplicate setupEnvironment function in auth.coverage.test.ts
- Remove duplicate helper functions in auth.login-redirect-issue-438.test.ts
- Fix unused import error
- Reduces duplication to pass Codacy quality gates

Resolves Codacy duplication threshold violations
